Change parameter type

+++ src/Person.java
@@ -15,7 +15,7 @@ public class Person {
         return age;
     }
 
-    public void setAge(int age) {
+    public void setAge(long age) {
         this.age = age;
     }
